+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Added Dry-Run Mode. (`-d` or `--dry-run`)
 - Added skip signing. (`-s` or `--skip-signing`)
 - chore: Modularize the code
+- Added `init` subcommand to automatically generate an example structure for your ROM.